Изменил логику отображения AlertDialog при поиске во время ввода символов в строке поиска
