+++ core/src/main/java/com/example/core/BaseActivity.kt
@@ -8,13 +8,16 @@ import com.elchaninov.utils.AlertDialogFragment
 import com.elchaninov.utils.convertMeaningsToString
 import com.elchaninov.utils.isOnline
 import com.example.core.databinding.LoadingFrameLayoutBinding
+import com.example.core.view.SearchDialogFragment
 import com.example.core.viewmodel.BaseViewModel
 
-abstract class BaseActivity<T : AppState> : AppCompatActivity() {
+abstract class BaseActivity<T : AppState> : AppCompatActivity(), SearchDialogFragment.OnSearchClickListener {
 
     private lateinit var binding: LoadingFrameLayoutBinding
     abstract val model: BaseViewModel<T>
     abstract fun renderData(appState: T)
+
+    private var isEnableShowErrorIfEmpty = true
 
     protected val isOnline: Boolean
         get() = isOnline(applicationContext)
@@ -50,7 +53,7 @@ abstract class BaseActivity<T : AppState> : AppCompatActivity() {
     }
 
     protected fun showAlertDialog(title: String?, message: String?, buttonTitle: String? = null) {
-        if (isDialogNull()) {
+        if (isDialogNull() && isEnableShowErrorIfEmpty) {
             AlertDialogFragment.newInstance(title, message, buttonTitle)
                 .show(supportFragmentManager, DIALOG_FRAGMENT_TAG)
         }
@@ -58,6 +61,16 @@ abstract class BaseActivity<T : AppState> : AppCompatActivity() {
 
     private fun isDialogNull(): Boolean {
         return supportFragmentManager.findFragmentByTag(DIALOG_FRAGMENT_TAG) == null
+    }
+
+    override fun onClick(searchWord: String) {
+        isEnableShowErrorIfEmpty = true
+        model.getData(searchWord, isOnline)
+    }
+
+    override fun onFlowSearch(searchWord: String) {
+        isEnableShowErrorIfEmpty = false
+        model.getData(searchWord, isOnline)
     }
 
     companion object {
